UseAuthentication, UseAuthorization, and add SellerId to the Product

+++ src/Services/ProductCatalog/ProductCatalog.API/Products/DeleteProduct/DeleteProductEndpoint.cs
@@ -1,6 +1,10 @@
-﻿using ProductCatalog.API.Endpoints;
+﻿using Microsoft.AspNetCore.Authorization;
+using ProductCatalog.API.Constants;
+using ProductCatalog.API.Endpoints;
+using ProductCatalog.API.Extentions;
 using ProductCatalog.API.Models;
 using ProductCatalog.API.Persistance;
+using System.Security.Claims;
 
 namespace ProductCatalog.API.Products.DeleteProduct;
 
@@ -11,9 +15,13 @@ public class DeleteProductEndpoint : IEndpoint
 {
     public void MapEndpoint(IEndpointRouteBuilder app)
     {
-        app.MapDelete("api/products/{id}", async (int id, AppDbContext context) =>
+        app.MapDelete("api/products/{id}", [Authorize(Roles = Roles.Seller)] async (int id, AppDbContext context, ClaimsPrincipal user) =>
         {
             var product = context.Products.FirstOrDefault(p => p.Id == id);
+
+            if (user.GetLoggedInUserId() != product?.SellerId)
+                return Results.BadRequest();
+
             if (product is null)
             {
                 var response = new DeleteProductResponse(false);
@@ -28,6 +36,7 @@ public class DeleteProductEndpoint : IEndpoint
                 return Results.Ok(response);
             }
         })
+         .RequireAuthorization()
          .ProducesProblem(StatusCodes.Status404NotFound)
          .Produces<DeleteProductResponse>(StatusCodes.Status200OK)
          .WithTags(nameof(Product));
